Lesson#2/HW/K8S_FullStack_Exercise/flask-app/requirements.txt
flask
fido2
telemetry
mysql-connector-python
mysql-connector